+++ test/test_utils/routing/router_test.dart
@@ -4,13 +4,13 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:github_search/presentation/components/router.dart';
 import 'package:github_search/presentation/pages/error/error_page.dart';
 import 'package:github_search/presentation/pages/repo/repo_index_page.dart';
+import 'package:github_search/utils/routing/router.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../test_utils/test_agent.dart';
+import '../test_agent.dart';
 
 class _MockPageRoute extends Mock implements PageRoute<dynamic> {}
